refactor: SM-51 magic-number 린트 규칙 해제로 인한 수정
